Transpose Updates to Avoid Sorting (#3)
* [WIP] Use CSC and work with transposed jacobians, to avoid sorting.

Basically the idea is that everytime you transpose a BCOO the indeces
are not sorted anymore. So bettter to always work in the transposed
world and defer the tranpose once I update the gradients, for which it
will be smaller and in dense format. For that to work I need to return
the jacobian transposed in BCOO with the indices sorted.

* Add unit tests for the tranposed spares jacobian.

* Allow specifying if tranpose the jacobian or not.

* Allow to specify a specific loss function.

* Add stacked encoder-decoder & Refactor

* Fix tanh of RNN

* Add GRU Cell